organized folders and naming conventions; not complete
