+++ src/components/CalendarView/CalendarMonthView.jsx
@@ -22,5 +22,3 @@ export const CalendarMonthView = ({ title }) => {
     </div>
   );
 };
-
-// export default CalendarMonthView;
