+++ src/components/CalendarView/StartingBalance/StartingBalanceEntryEdit.jsx
@@ -3,7 +3,7 @@ import { formatCurrency, toCents } from "../../../services/currencyServices";
 
 import { Icon } from "../../shared/Icon";
 
-import "../../styles/CurrentMonthViewHeader.scss";
+import "../../../styles/CurrentMonthHeader.scss";
 
 export const StartingBalanceEntryEdit = ({
   startingBal,
