+++ src/components/LandingView/LandingView.jsx
@@ -4,20 +4,15 @@ import { SiteNavBar } from "../shared/HeadersFooters/SiteNavBar";
 import { LandingViewHero } from "./LandingViewHero";
 import { GetStarted } from "./GetStarted";
 import { PrivacyNotice } from "../shared/HeadersFooters/PrivacyNotice";
-import { ContentSideBar } from "../shared/HeadersFooters/ContentSideBar";
 import { SiteFooter } from "../shared/HeadersFooters/SiteFooter";
 
 import "../../styles/LandingView.scss";
 
-// Landing View
 export const LandingView = () => (
-  <section className="landing">
+  <section id="landing">
     <SiteNavBar />
     <LandingViewHero />
-    <div className="wrapper-content-wrapper">
-      <GetStarted />
-      <ContentSideBar />
-    </div>
+    <GetStarted />
     <PrivacyNotice />
     <SiteFooter />
   </section>
