+++ src/styles/App.scss
@@ -20,12 +20,10 @@ h1 {
 
 h2 {
   font-size: 22px;
-  font-weight: 300;
 }
 
 h3 {
   font-size: 18px;
-  font-weight: 300;
 }
 
 .cash-available {
